Consolidate HumanController and ProfileController under /Profile routes (#312)
Merge all profile-related actions from HumanController and ProfileController
into a unified ProfileController under /Profile routes. Own-profile actions
live at /Profile/Me/*, viewing other profiles at /Profile/{id}, admin actions
at /Profile/{id}/Admin/*, admin list at /Profile/Admin.

- Move contact management to dedicated ContactsController at /Contacts
- Move people search from TeamController to ProfileController at /Profile/Search
- Move API search from /api/humans/search to /api/profiles/search
- Add permanent redirects from all old /Human/* routes
- Delete HumanController and HumanApiController
- Update all views, tag helpers, JS, and cross-controller references
- Fix broken /Human/{id}/Picture URLs in ShiftsController, VolController,
  ShiftAdmin views (Profile/Picture was already the correct endpoint)
- Update section invariant doc with new routing table
