+++ src/Humans.Web/Authorization/MembershipRequiredFilter.cs
@@ -21,7 +21,9 @@ public class MembershipRequiredFilter : IAsyncActionFilter
         "Profile",     // Set up profile during onboarding
         "Admin",       // Has its own Roles = "Admin" gate
         "Board",       // Has its own Roles = "Board,Admin" gate
-        "Human",       // Public profile viewing
+        "Contacts",    // Admin contacts (has its own role gate)
+        "HumanRedirect",    // Old route redirects
+        "HumanApiRedirect", // Old API route redirects
         "Language",         // Language switching
         "OnboardingReview", // Has its own coordinator/Board role gate
         "Camp",             // Public camps pages ([AllowAnonymous])
