+++ src/Humans.Web/Controllers/VolController.cs
@@ -210,7 +210,7 @@ public class VolController : HumansControllerBase
                                             RemainingSlots = u.RemainingSlots,
                                             Signups = u.Signups.Select(s => new ShiftSignupInfo(
                                                 s.UserId, s.DisplayName, s.Status,
-                                                s.HasProfilePicture ? $"/Human/{s.UserId}/Picture" : null))
+                                                s.HasProfilePicture ? $"/Profile/Picture?id={s.UserId}" : null))
                                                 .ToList()
                                         };
                                     }).OrderBy(s => s.AbsoluteStart).ToList()
