+++ src/Humans.Web/Views/Contacts/Create.cshtml
@@ -4,7 +4,7 @@
 }
 
 <div class="mb-3">
-    <a asp-action="Contacts" class="btn btn-outline-secondary btn-sm">
+    <a asp-action="Index" class="btn btn-outline-secondary btn-sm">
         <i class="fa-solid fa-arrow-left"></i> Back to Contacts
     </a>
 </div>
@@ -14,7 +14,7 @@
 
 <div class="card" style="max-width: 600px;">
     <div class="card-body">
-        <form method="post" asp-action="CreateContact">
+        <form method="post" asp-action="Create">
             <div asp-validation-summary="All" class="text-danger mb-3"></div>
 
             <div class="mb-3">
