+++ src/Humans.Web/Views/Contacts/Detail.cshtml
@@ -4,7 +4,7 @@
 }
 
 <div class="mb-3">
-    <a asp-action="Contacts" class="btn btn-outline-secondary btn-sm">
+    <a asp-action="Index" class="btn btn-outline-secondary btn-sm">
         <i class="fa-solid fa-arrow-left"></i> Back to Contacts
     </a>
 </div>
