+++ src/Humans.Web/Views/Profile/AdminDetail.cshtml
@@ -10,8 +10,8 @@
 
 <nav aria-label="breadcrumb">
     <ol class="breadcrumb">
-        <li class="breadcrumb-item"><a asp-action="Humans">Humans</a></li>
-        <li class="breadcrumb-item"><a asp-action="HumanProfile" asp-route-id="@Model.UserId">@Model.DisplayName</a></li>
+        <li class="breadcrumb-item"><a asp-action="AdminList">Humans</a></li>
+        <li class="breadcrumb-item"><a asp-action="ViewProfile" asp-route-id="@Model.UserId">@Model.DisplayName</a></li>
         <li class="breadcrumb-item active" aria-current="page">Admin</li>
     </ol>
 </nav>
@@ -285,7 +285,7 @@
         <div class="card mb-4">
             <div class="card-body d-flex justify-content-between align-items-center">
                 <span><i class="fa-solid fa-envelope me-2"></i> Email Outbox</span>
-                <a asp-action="Outbox" asp-route-id="@Model.UserId" class="btn btn-sm btn-outline-primary">
+                <a asp-action="AdminOutbox" asp-route-id="@Model.UserId" class="btn btn-sm btn-outline-primary">
                     View emails <span class="badge bg-secondary ms-1">@outboxCount</span>
                 </a>
             </div>
