+++ src/Humans.Web/Views/Profile/Index.cshtml
@@ -160,7 +160,7 @@
                     <a asp-controller="Profile" asp-action="Notifications" class="list-group-item list-group-item-action">Notifications</a>
                     <a asp-controller="Governance" asp-action="Index" class="list-group-item list-group-item-action">@Localizer["Nav_Governance"]</a>
                     <a asp-controller="Profile" asp-action="Privacy" class="list-group-item list-group-item-action">@Localizer["ProfilePrivacy_Title"]</a>
-                    <a asp-controller="Profile" asp-action="Outbox" class="list-group-item list-group-item-action">My Emails</a>
+                    <a asp-controller="Profile" asp-action="MyOutbox" class="list-group-item list-group-item-action">My Emails</a>
                 </div>
             </div>
         </div>
