+++ src/Humans.Web/Views/Profile/Search.cshtml
@@ -6,7 +6,7 @@
 
 <nav aria-label="breadcrumb">
     <ol class="breadcrumb">
-        <li class="breadcrumb-item"><a asp-action="Index">@Localizer["Teams_Title"]</a></li>
+        <li class="breadcrumb-item"><a asp-controller="Team" asp-action="Index">@Localizer["Teams_Title"]</a></li>
         <li class="breadcrumb-item active">@Localizer["Search_Title"]</li>
     </ol>
 </nav>
