+++ src/Humans.Web/Views/Shared/_Layout.cshtml
@@ -88,7 +88,7 @@
                         @if (Humans.Web.Authorization.RoleChecks.IsHumanAdmin(User) && !Humans.Web.Authorization.RoleChecks.IsAdminOrBoard(User))
                         {
                             <li class="nav-item">
-                                <a class="nav-link" asp-area="" asp-controller="Human" asp-action="Humans">Humans</a>
+                                <a class="nav-link" asp-area="" asp-controller="Profile" asp-action="AdminList">Humans</a>
                             </li>
                         }
                         @if (Humans.Web.Authorization.RoleChecks.IsAdmin(User))
