+++ src/Humans.Web/Views/Team/Index.cshtml
@@ -11,7 +11,7 @@
         <div class="d-flex gap-2">
             <vc:access-matrix section="Teams" />
             <a asp-action="MyTeams" class="btn btn-outline-secondary">@Localizer["MyTeams_Title"]</a>
-            <a asp-action="Search" class="btn btn-outline-secondary"><i class="fa-solid fa-search"></i></a>
+            <a asp-controller="Profile" asp-action="Search" class="btn btn-outline-secondary"><i class="fa-solid fa-search"></i></a>
             <a asp-action="Roster" class="btn btn-outline-secondary">@Localizer["Nav_Roster"]</a>
             <a asp-action="Birthdays" class="btn btn-outline-secondary">@Localizer["Birthdays_Title"]</a>
             <a asp-action="Map" class="btn btn-outline-secondary">@Localizer["Map_Title"]</a>
